Implementado Cadastro de Horários de Médicos

+++ src/controller/medicoController/excluir_med.php
@@ -6,7 +6,7 @@ define('HOME_DIR', 'C:\xampp\htdocs\modelo-fc\src\model');
 include_once HOME_DIR . '\config-banco-dados.php';
 
 $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
-// if(!empty($id)){
+if(!empty($id)){
 	$exc = "DELETE FROM medicos WHERE id='$id'";
     
     $exclusao = $conn->prepare($exc);
@@ -17,15 +17,8 @@ $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
         //redirecionamento
         header("Location: ../index.php");
     } 
-// 	if(mysqli_affected_rows($conn)){
-// 		$_SESSION['msg'] = "<p style='color:green;'>Usuário apagado com sucesso</p>";
-// 		header("Location: index.php");
-// 	}else{
-		
-// 		$_SESSION['msg'] = "<p style='color:red;'>Erro o usuário não foi apagado com sucesso</p>";
-// 		header("Location: index.php");
-// 	}
-// }else{	
-// 	$_SESSION['msg'] = "<p style='color:red;'>Necessário selecionar um usuário</p>";
-// 	header("Location: index.php");
-// }
+
+}else{	
+	$_SESSION['msg'] = "<p style='color:red;'>Necessário selecionar um usuário</p>";
+	header("Location: ../index.php");
+}
